% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.tsae.R
\name{ecm.tsae}
\alias{ecm.tsae}
\title{EBLUPs univariate small area estimation using t-distrubiton approach}
\usage{
ecm.tsae(
  formula,
  vardir,
  sigma2u,
  df = 3,
  maxiter = 1000,
  tolerance = 10^(-4),
  data
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} that describe the fitted model.}

\item{vardir}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{sigma2u}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{df}{number}

\item{maxiter}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{tolerance}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{data}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}
}
\description{
This function is used for small area estimation thats have outlier in respons variable
}
