% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.tsae.R
\name{ecm.tsae}
\alias{ecm.tsae}
\title{Parametric Bootstrap Mean Squared Error of EBLUPs based on a Univariate Fay Herriot model with Additive Logistic Transformation for Non-Sampled Data}
\usage{
ecm.tsae(
  formula,
  vardir,
  sigma2u,
  df = 3,
  maxiter = 1000,
  tolerance = 10^(-4),
  data
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} that describe the fitted model.}

\item{vardir}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{sigma2u}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{df}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{maxiter}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{tolerance}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{data}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}
}
\description{
This function gives the MSE of transformed EBLUP based on a univariate Fay-Herriot model. For sampled domains, MSE is estimated using modified parametric bootstrap approach proposed by Butar & Lahiri. For non-sampled domains, MSE is estimated using modified approach proposed by Haris & Ubaidillah.
}
