% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.tsae.R
\name{ecm.tsae}
\alias{ecm.tsae}
\title{EBLUPs univariate small area estimation using t-distrubiton approach}
\usage{
ecm.tsae(
  formula,
  vardir,
  sigma2u,
  df = 3,
  maxiter = 1000,
  tolerance = 10^(-4),
  data
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} that describe the fitted model.}

\item{vardir}{sampling variances of direct estimations. If data is defined, it is a vector containing names of sampling variance columns.}

\item{sigma2u}{number random for first sigma2u, Default \code{0}.}

\item{df}{number containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables, Default:\code{3}.}

\item{maxiter}{maximum number of iterations allowed in the Fisher-scoring algorithm, Default: \code{1000}.}

\item{tolerance}{convergence tolerance limit for the Fisher-scoring algorithm, Default: \code{1e-4}.}

\item{data}{matrix that include vardir, auxiliary variable, response variabel.}
}
\description{
This function is used for small area estimation thats have outlier in respons variable
}
