% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.tsae.R
\name{mse.tsae}
\alias{mse.tsae}
\title{Parametric Mean Squared Error of EBLUPs based on a Univariate small area estimation using t-distribution approach}
\usage{
mse.tsae(formula, sigma2u.awal, vardir, df, maxiter, tolerance, data)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} that describe the fitted model.}

\item{sigma2u.awal}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{vardir}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{df}{Scalar containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{maxiter}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{tolerance}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}

\item{data}{vector containing the sampling variances of direct estimators for each domain. The values must be sorted as the variables in \code{formula}.}
}
\description{
This function gives MSE based on univariate t-distribution model
}
